### Corrections:

- Comparison operator like `-g` or `>` only valid for _integer_ doesn't
not require
  quote but still works fine if we add quote.

- Same as above but we should treat _strings_ with quote.

> Yes we can add quote (`""`) on _integer_ and without quote on
_strings_
but this only for good practices.

### Features:

This only minor feat which we can handle user input properly.

### Breaking Changes?
No

+++ helium.sh
@@ -20,7 +20,7 @@ function header() {
 }
 
 function isRoot() {
-	if [ "${EUID}" != 0 ]; then
+	if [ ${EUID} != 0 ]; then
 		echo " You need to run this script as root"
 		exit 1
 	fi
@@ -40,11 +40,11 @@ function checkVirt() {
 function checkOS() {
 	if [[ $(grep -w "ID" /etc/os-release | awk -F'=' '{print $2}') -ne "debian" ]] || [[ $(grep -w "ID" /etc/os-release | awk -F'=' '{print $2}') -ne "ubuntu" ]]; then
 		clear
-        	header
- 		echo
- 		echo -e ${RED}" Your OS is not supported. Please use Debian/Ubuntu"$NOCOLOR
- 		echo ""
- 		exit 1
+		header
+		echo
+		echo -e ${RED}" Your OS is not supported. Please use Debian/Ubuntu"$NOCOLOR
+		echo ""
+		exit 1
 	fi
 }
 
@@ -58,27 +58,27 @@ function install() {
 	echo -e " Installing Helium..."
 	[[ ! -e /etc/dnsmasq ]] && mkdir -p /etc/dnsmasq
 	[[ ! -e /etc/resolv.conf.bak ]] && cp /etc/resolv.conf /etc/resolv.conf.bak
-    	if [[ $(lsof -i :53 | grep -w -c "systemd-r") -ge 1 ]]; then
-    		systemctl disable systemd-resolved
+	if [[ $(lsof -i :53 | grep -w -c "systemd-r") -ge 1 ]]; then
+		systemctl disable systemd-resolved
 		systemctl stop systemd-resolved
 		unlink /etc/resolv.conf
-		echo "nameserver 1.1.1.1" > /etc/resolv.conf
-    	fi
-    	apt update && apt install -y dnsmasq dnsutils vnstat resolvconf
+		echo "nameserver 1.1.1.1" >/etc/resolv.conf
+	fi
+	apt update && apt install -y dnsmasq dnsutils vnstat resolvconf
 	systemctl enable dnsmasq
-    	mv /etc/dnsmasq.conf /etc/dnsmasq.conf.bak
+	mv /etc/dnsmasq.conf /etc/dnsmasq.conf.bak
 	rm -rf /etc/dnsmasq.conf
-    	wget -q -O /etc/dnsmasq.conf "https://raw.githubusercontent.com/abidarwish/helium/main/dnsmasq.conf"
-    	sed -i "s/YourPublicIP/${publicIP}/" /etc/dnsmasq.conf
+	wget -q -O /etc/dnsmasq.conf "https://raw.githubusercontent.com/abidarwish/helium/main/dnsmasq.conf"
+	sed -i "s/YourPublicIP/${publicIP}/" /etc/dnsmasq.conf
 	rm -rf ${providers}
-    	wget -q -O ${providers} "https://raw.githubusercontent.com/abidarwish/helium/main/providers.txt"
+	wget -q -O ${providers} "https://raw.githubusercontent.com/abidarwish/helium/main/providers.txt"
 	sleep 1
 	updateEngine
-    	>/etc/resolvconf/resolv.conf.d/original
+	>/etc/resolvconf/resolv.conf.d/original
 	echo "nameserver 127.0.0.1" >/etc/resolv.conf
-    	echo "nameserver 127.0.0.1" >/etc/resolvconf/resolv.conf.d/head
+	echo "nameserver 127.0.0.1" >/etc/resolvconf/resolv.conf.d/head
 	sleep 1
-    	echo -e " Installation completed"
+	echo -e " Installation completed"
 	sleep 1
 	echo -e " Type \e[1;32mhelium\e[0m to start"
 	echo
@@ -113,7 +113,7 @@ function stop() {
 		echo -e $GREEN" Helium is running"$NOCOLOR
 		echo
 		read -p " Do you want to stop Helium? [y/n]: " STOP
-		if [[ ${STOP} == "y" ]]; then
+		if [[ ${STOP,,} == "y" ]]; then
 			systemctl disable dnsmasq
 			systemctl stop dnsmasq
 			echo "nameserver 1.1.1.1" >/etc/resolv.conf
@@ -136,21 +136,21 @@ function stop() {
 
 function changeDNS() {
 	clear
-   	header
-    	echo
-    	echo -e " Proxy server IP address to bypass Netflix"
-    	read -p " (press c to cancel): " DNS
-    	oldDNS=$(grep -E -w "^server" /etc/dnsmasq.conf | cut -d '=' -f2)
-    	[[ ${DNS} == "c" ]] && mainMenu
-    	[[ -z ${DNS} ]] && changeDNS
-    	sed -i "s/server=${oldDNS}/server=${DNS}/" /etc/dnsmasq.conf
-        systemctl restart dnsmasq
-    	sleep 1
-    	echo -e -n " DNS server has been changed to "
-    	echo -e $GREEN"$DNS"$NOCOLOR
-    	echo
-    	read -p " Press Enter to continue..."
-    	mainMenu
+	header
+	echo
+	echo -e " Proxy server IP address to bypass Netflix"
+	read -p " (press c to cancel): " DNS
+	oldDNS=$(grep -E -w "^server" /etc/dnsmasq.conf | cut -d '=' -f2)
+	[[ ${DNS,,} == "c" ]] && mainMenu
+	[[ -z $DNS ]] && changeDNS
+	sed -i "s/server=${oldDNS}/server=${DNS}/" /etc/dnsmasq.conf
+	systemctl restart dnsmasq
+	sleep 1
+	echo -e -n " DNS server has been changed to "
+	echo -e $GREEN"$DNS"$NOCOLOR
+	echo
+	read -p " Press Enter to continue..."
+	mainMenu
 }
 
 function reinstall() {
@@ -158,11 +158,11 @@ function reinstall() {
 	header
 	echo
 	read -p " Do you want to reinstall Helium? [y/n]: " REINSTALL
-	[[ ${REINSTALL} != "y" ]] && mainMenu
+	[[ ${REINSTALL,,} != "y" ]] && mainMenu
 	echo -e " Reinstalling Helium..."
 	sleep 2
 	[[ ! -e /etc/dnsmasq ]] && mkdir -p /etc/dnsmasq
-    	echo "nameserver 1.1.1.1" > /etc/resolv.conf
+	echo "nameserver 1.1.1.1" >/etc/resolv.conf
 	apt update && apt install -y dnsmasq dnsutils vnstat resolvconf
 	systemctl enable dnsmasq
 	rm -rf /etc/dnsmasq.conf
@@ -187,31 +187,38 @@ function uninstall() {
 	header
 	echo
 	read -p " Do you want to uninstall Helium? [y/n]: " UNINSTALL
-	[[ ${UNINSTALL} != "y" ]] && mainMenu
-	systemctl stop dnsmasq
-	systemctl disable dnsmasq
-	apt remove -y dnsmasq > /dev/null 2>&1
-	rm -rf /etc/dnsmasq.d
-	rm -rf /etc/dnsmasq
-	>/etc/resolvconf/resolv.conf.d/original
-	>/etc/resolvconf/resolv.conf.d/head
-	mv /etc/resolv.conf.bak /etc/resolv.conf
-	echo -e -n " Uninstalling Helium..."
-	sleep 2
-	echo -e $GREEN"done"$NOCOLOR
-	echo
-	exit 0
+	if [[ ${UNINSTALL,,} == "y" ]]; then
+		systemctl stop dnsmasq
+		systemctl disable dnsmasq
+		apt remove -y dnsmasq >/dev/null 2>&1
+		rm -rf /etc/dnsmasq.d
+		rm -rf /etc/dnsmasq
+		>/etc/resolvconf/resolv.conf.d/original
+		>/etc/resolvconf/resolv.conf.d/head
+		mv /etc/resolv.conf.bak /etc/resolv.conf
+		echo -e -n " Uninstalling Helium..."
+		sleep 2
+		echo -e $GREEN"done"$NOCOLOR
+		echo
+		exit 0
+	else
+		mainMenu
+	fi
 }
 
 function updateEngine() {
 	echo -e -n " Updating blocked hostnames..."
 	>${tempHostsList}
 	while IFS= read -r line; do
-        list_url=$(echo $line | grep -E -v "^#" | cut -d '"' -f2)
-        curl "${list_url}" 2> /dev/null | sed -E '/^!/d' | sed '/#/d' | sed -E 's/^\|\|/0.0.0.0 /g' | awk -F '^' '{print $1}' | grep -E "^0.0.0.0" >> ${tempHostsList}
+		list_url=$(echo $line | grep -E -v "^#" | cut -d '"' -f2)
+		curl "${list_url}" 2>/dev/null | sed -E '/^!/d' | sed '/#/d' | sed -E 's/^\|\|/0.0.0.0 /g' | awk -F '^' '{print $1}' | grep -E "^0.0.0.0" >>${tempHostsList}
 	done <${providers}
 	if [[ ! -z $(ip a | grep -w "inet6") ]]; then
 		grep -E "^0.0.0.0" ${tempHostsList} | sed -E 's/^0.0.0.0/::1/g' >>${tempHostsList}
+	fi
+	cat ${tempHostsList} | sed '/^$/d' | sed -E '/^0.0.0.0 0.0.0.0/d; /^::1 0.0.0.0/d' | sort | uniq >${dnsmasqHostFinalList}
+	if [[ ! -e /etc/dnsmasq/whitelist.hosts ]]; then
+		touch /etc/dnsmasq/whitelist.hosts
 	fi
 	cat ${tempHostsList} | sed '/^$/d' | sed -E '/^0.0.0.0 0.0.0.0/d' | sed -E '/^::1 0.0.0.0/d' | sort | uniq >${dnsmasqHostFinalList}
 	[[ ! -e /etc/dnsmasq/whitelist.hosts ]] && touch /etc/dnsmasq/whitelist.hosts
@@ -222,8 +229,8 @@ function updateEngine() {
 	systemctl restart dnsmasq
 	echo -e ${GREEN}"done"${NOCOLOR}
 	sleep 1
-	echo -e -n ${GREEN}" $(cat ${dnsmasqHostFinalList} | sed '/^$/d' | wc -l) "${NOCOLOR}
-   	echo -e "hostnames have been blocked"
+	echo -e -n $GREEN" $(cat ${dnsmasqHostFinalList} | sed '/^$/d' | wc -l) "$NOCOLOR
+	echo -e "hostnames have been blocked"
 }
 
 function listUpdate() {
@@ -231,11 +238,14 @@ function listUpdate() {
 	header
 	echo
 	read -p " Do you want to update blocked hostnames? [y/n]: " UPDATE
-	[[ ${UPDATE} != "y" ]] && mainMenu
-	updateEngine
-	echo
-	read -p " Press Enter to continue..."
-	mainMenu
+	if [[ ${UPDATE,,} == "y" ]]; then
+		updateEngine
+		echo
+		read -p " Press Enter to continue..."
+		mainMenu
+	else
+		mainMenu
+	fi
 }
 
 function activateProvider() {
@@ -249,7 +259,7 @@ function activateProvider() {
 		ACTIVE_PROVIDER=$(echo $line | grep -v -E "^#" | cut -d '=' -f1)
 		INACTIVE_PROVIDER=$(echo $line | grep -E "^#" | cut -d '=' -f1 | sed -E 's/^#//g')
 		if [[ $(echo $line | grep -v -c -E "^#") -gt 0 ]]; then
-		 	printf " %-25s \e[1;32m%12s\e[0m\n" "${ACTIVE_PROVIDER}" "active"
+			printf " %-25s \e[1;32m%12s\e[0m\n" "${ACTIVE_PROVIDER}" "active"
 		else
 			printf " %-25s \e[1;31m%12s\e[0m\n" "${INACTIVE_PROVIDER}" "inactive"
 		fi
@@ -262,27 +272,27 @@ function activateProvider() {
 		read -p " Select a provider to be activated
  (press c to cancel): " SELECT
 	fi
-    	if [[ ${SELECT} == "s" ]]; then
+	if [[ ${SELECT,,} == "s" ]]; then
 		mv /etc/dnsmasq/providers.tmp /etc/dnsmasq/providers.txt
 		echo " Applying changes..."
 		updateEngine
 		echo
 		read -p " Press Enter to continue..."
 		mainMenu
-    	fi
-    	if [[ ${SELECT} == "c" ]]; then
+	fi
+	if [[ ${SELECT,,} == "c" ]]; then
 		rm -rf /etc/dnsmasq/providers.tmp
 		mainMenu
 	fi
-	[[ -z ${SELECT} ]] && activateProvider
+	[[ -z $SELECT ]] && activateProvider
 	if [[ $(grep -E -c -w "^#${SELECT}" /etc/dnsmasq/providers.tmp) != 0 ]]; then
 		sed -E -i "s/^\#${SELECT}/${SELECT}/" /etc/dnsmasq/providers.tmp
 		activateProvider
 	else
 		echo -e " ${SELECT} is already active"
 	fi
-    	echo
-    	read -p " Press Enter to continue..."
+	echo
+	read -p " Press Enter to continue..."
 	activateProvider
 }
 
@@ -297,7 +307,7 @@ function deactivateProvider() {
 		ACTIVE_PROVIDER=$(echo $line | grep -v -E "^#" | cut -d '=' -f1)
 		INACTIVE_PROVIDER=$(echo $line | grep -E "^#" | cut -d '=' -f1 | sed -E 's/^#//g')
 		if [[ $(echo $line | grep -v -c -E "^#") -gt 0 ]]; then
-		 	printf " %-25s \e[1;32m%12s\e[0m\n" "${ACTIVE_PROVIDER}" "active"
+			printf " %-25s \e[1;32m%12s\e[0m\n" "${ACTIVE_PROVIDER}" "active"
 		else
 			printf " %-25s \e[1;31m%12s\e[0m\n" "${INACTIVE_PROVIDER}" "inactive"
 		fi
@@ -307,10 +317,10 @@ function deactivateProvider() {
 		read -p " Select a provider to be activated
  (press s to save changes or c to cancel): " SELECT
 	else
-       	read -p " Select a provider to be deactivated
+		read -p " Select a provider to be deactivated
  (press c to cancel): " SELECT
 	fi
-	if [[ ${SELECT} == "s" ]]; then
+	if [[ ${SELECT,,} == "s" ]]; then
 		mv /etc/dnsmasq/providers.tmp /etc/dnsmasq/providers.txt
 		echo " Applying changes..."
 		updateEngine
@@ -318,11 +328,11 @@ function deactivateProvider() {
 		read -p " Press Enter to continue..."
 		mainMenu
 	fi
-	if [[ ${SELECT} == "c" ]]; then
+	if [[ ${SELECT,,} == "c" ]]; then
 		rm -rf /etc/dnsmasq/providers.tmp
 		mainMenu
 	fi
-	[[ -z ${SELECT} ]] && deactivateProvider
+	[[ -z $SELECT ]] && deactivateProvider
 	if [[ $(grep -E -c -w "^#${SELECT}" /etc/dnsmasq/providers.tmp) == 0 ]]; then
 		sed -E -i "s/^${SELECT}/\#${SELECT}/" /etc/dnsmasq/providers.tmp
 		deactivateProvider
@@ -352,17 +362,17 @@ function whitelistHost() {
 		read -p " Select a url from above to delete or type a new one to whitelist
  (press s to save changes or c to cancel): " SELECT
 	else
-       	read -p " Select a url from above to delete or type a new one to whitelist
+		read -p " Select a url from above to delete or type a new one to whitelist
  (press c to cancel): " SELECT
 	fi
-	if [[ ${SELECT} == "s" ]]; then
-       	mv /etc/dnsmasq/whitelist.hosts.tmp /etc/dnsmasq/whitelist.hosts
+	if [[ ${SELECT,,} == "s" ]]; then
+		mv /etc/dnsmasq/whitelist.hosts.tmp /etc/dnsmasq/whitelist.hosts
 		updateEngine
 		echo
 		read -p " Press Enter to continue..."
 		mainMenu
 	fi
-	if [[ ${SELECT} == "c" ]]; then
+	if [[ ${SELECT,,} == "c" ]]; then
 		rm -rf /etc/dnsmasq/whitelist.hosts.tmp
 		mainMenu
 	fi
@@ -373,7 +383,7 @@ function whitelistHost() {
 		whitelistHost
 	else
 		read -p " Do you want to delete this url? [y/n]: " DELETE
-		if [[ ${DELETE} == "y" ]]; then
+		if [[ ${DELETE,,} == "y" ]]; then
 			sed -E -i "/^${SELECT}/d" /etc/dnsmasq/whitelist.hosts.tmp
 			whitelistHost
 		else
@@ -387,7 +397,7 @@ function cleaner() {
 	header
 	echo
 	read -p " Do you want to cleanup the database? [y/n]: " CLEANUP
-	[[ ${CLEANUP} != "y" ]] && mainMenu
+	[[ ${CLEANUP,,} != "y" ]] && mainMenu
 	ORIGINAL_DATABASE=$(cat /etc/dnsmasq/adblock.hosts | sed '/^$/d' | wc -l)
 	echo -e -n " Checking database..."
 	sleep 2
@@ -400,7 +410,7 @@ function cleaner() {
 		echo -e -n ${RED}"deleted"${NOCOLOR}
 	done
 	NEW_DATABASE=$(cat /etc/dnsmasq/adblock.hosts | sed '/^$/d' | wc -l)
-	DELETED_HOSTNAMES=$(( ORIGINAL_DATABASE-NEW_DATABASE ))
+	DELETED_HOSTNAMES=$((ORIGINAL_DATABASE - NEW_DATABASE))
 	printf "\n \e[1;32m%'d\e[0m %-10s\n" "${DELETED_HOSTNAMES}" "dead hostnames have been deleted from the database"
 	[[ ${DELETED_HOSTNAMES} -gt 0 ]] && systemctl restart dnsmasq
 	echo
@@ -433,13 +443,13 @@ function updateHelium() {
 	echo
 	echo
 	read -p " Do you want to update? [y/n]: " UPDATE
-	if [[ ${UPDATE} != "y" ]]; then
+	if [[ ${UPDATE,,} != "y" ]]; then
 		rm -rf /tmp/helium.tmp
 		mainMenu
 	fi
 	read -p " Do you want to overwrite the existing providers? [y/n]: " OVERWRITE
 	echo -n -e " Updating Helium..."
-	if [[ ${OVERWRITE} == "y" ]]; then
+	if [[ ${OVERWRITE,,} == "y" ]]; then
 		rm -rf /etc/dnsmasq/providers.txt
 		wget -q -O /etc/dnsmasq/providers.txt "https://raw.githubusercontent.com/abidarwish/helium/main/providers.txt"
 	fi
@@ -465,8 +475,8 @@ function mainMenu() {
 	header
 	echo
 	echo -e " \e[1mSystem Status\e[0m"
-	if [[ $(systemctl is-active dnsmasq) == active ]]; then
-        printf " %-25s %1s \e[1;32m%7s\e[0m" "Dnsmasq" ":" "running"
+	if [[ $(systemctl is-active dnsmasq) == "active" ]]; then
+		printf " %-25s %1s \e[1;32m%7s\e[0m" "Dnsmasq" ":" "running"
 		printf "\n %-25s %1s \e[1;32m%7s\e[0m" "Active since" ":" "$(systemctl status dnsmasq.service | grep -w "Active" | awk '{print $9,$10,$11,$12}')"
 	else
 		printf " %-25s %1s \e[1;31m%7s\e[0m" "Dnsmasq" ":" "stopped"
@@ -474,12 +484,12 @@ function mainMenu() {
 	NAMESERVER=$(grep -w -E "^server" /etc/dnsmasq.conf | head -n 1 | awk -F'=' '{print $2}')
 	printf "\n %-25s %1s \e[1;32m%7s\e[0m" "Nameserver" ":" "$NAMESERVER"
 	printf "\n %-25s %1s \e[1;32m%'d\n\e[0m" "Blocked hostnames" ":" "$(cat ${dnsmasqHostFinalList} | wc -l)"
-	echo
+  echo
 	echo -e " \e[1mMachine Info\e[0m"
 	CPU=$(cat /proc/cpuinfo | grep "model\|Model" | tail -n 1 | awk -F: '{print $2}' | cut -d " " -f2-4)
 	CPU_CORE=$(lscpu | grep "CPU(s)" | head -n 1 | awk '{print $2}')
 	CPU_MHZ=$(lscpu | grep "MHz" | head -n 1 | sed 's/ //g' | awk -F: '{print $2}' | cut -d. -f1)
-	OS=$(cat /etc/os-release |grep "PRETTY_NAME" | awk -F\" '{print $2}')
+	OS=$(cat /etc/os-release | grep "PRETTY_NAME" | awk -F\" '{print $2}')
 	KERNEL=$(uname -r)
 	RAM_USED=$(free -m | grep Mem: | awk '{print $3}')
 	TOTAL_RAM=$(free -m | grep Mem: | awk '{print $2}')
@@ -498,13 +508,13 @@ function mainMenu() {
 	printf "\n %-25s %1s %-7s\e[0m" "RAM Usage" ":" "${RAM_USED}MB / ${TOTAL_RAM}MB (${RAM_USAGE}%)"
 	printf "\n %-25s %1s %-7s\e[0m" "Date" ":" "${DATE}"
 	printf "\n %-25s %1s %-7s\e[0m" "Uptime" ":" "${UPTIME}"
- 	printf "\n %-25s %1s %-7s\e[0m" "IP Address" ":" "${publicIP}"
+	printf "\n %-25s %1s %-7s\e[0m" "IP Address" ":" "${publicIP}"
 	printf "\n %-25s %1s %-7s\e[0m" "Daily Data Usage" ":" "${DAILY_USAGE}"
 	printf "\n %-25s %1s %-7s\e[0m" "Monthly Data Usage" ":" "${MONTHLY_USAGE}"
 	echo
 	echo
 	echo -e $WHITE" Manage Helium"$NOCOLOR
- 	echo -e " [01] Start Dnsmasq\t   [07] Whitelist host
+	echo -e " [01] Start Dnsmasq\t   [07] Whitelist host
  [02] Stop Dnsmasq\t   [08] Bypass Netflix
  [03] Update database\t   [09] Update Helium
  [04] Cleanup database\t   [10] Reinstall Helium
@@ -513,33 +523,33 @@ function mainMenu() {
 	echo
 	read -p $' Enter option [1-12]: ' MENU_OPTION
 	case ${MENU_OPTION} in
-	01|1)
+	01 | 1)
 		start
-	   	;;
-	02|2)
+		;;
+	02 | 2)
 		stop
 		;;
-	03|3)
+	03 | 3)
 		listUpdate
 		;;
-	04|4)
+	04 | 4)
 		cleaner
 		;;
-   	05|5)
+	05 | 5)
 		activateProvider
 		;;
-    	06|6)
-        	deactivateProvider
-        	;;
-	07|7)
+	06 | 6)
+		deactivateProvider
+		;;
+	07 | 7)
 		whitelistHost
 		;;
-	08|8)
-	       changeDNS
-	       ;;
-	09|9)
-	       updateHelium
-	       ;;
+	08 | 8)
+		changeDNS
+		;;
+	09 | 9)
+		updateHelium
+		;;
 	10)
 		reinstall
 		;;
@@ -547,10 +557,11 @@ function mainMenu() {
 		uninstall
 		;;
 	12)
-	       exit 0
+		exit 0
 		;;
 	*)
-	mainMenu
+		mainMenu
+		;;
 	esac
 }
